Don't remove old entries when loading calendars
The /load-calendar link used to remove all existing event entries.  Now
it only removes existing entries that are in the same time range as the
ones that have been found in the newly-parsed files.

This makes it possible to read next year's OGD files without killing the
current year's entries.  A new index (by type and date) had to be added
to the datastore so that we can selectively drop entries of a time
range.  (It's a bit unfortunate that ERZ only publishes data of a single
year on its OGD site.)

In the process of making this change, the entire process of parsing the
calendar files was refactored into a new ParsedAbholCSV class.

A problem with UTF-8 parsing was fixed at the same time, so that Umlauts
should now be represented correctly in the data store, and hopefully
also in calendar entries.