Introduce "year" as (optional) parameter to /load-calendar
The /load-calendar handler now takes the year as an argument.  This
defaults to the current year, or the next year when called late in the
year, i.e. in November or December.

The encoding of the URL pattern used since 2022 has been simplified so
that the same pattern can be used independently of the year, as long
as data.stadt-zuerich.ch sticks to their own convention.  Ideally,
this will make it unnecessary to modify this code in the future just
to load next year's calendars.

Also, diagnostic messages have been improved a bit when loading a
calendar fails.