Obfuscate load-calendar URL
It seemed to be hit by some crawlers.  This was good in the sense that
the calendar would be updated automagically, and I didn't have to
think about it at the end of every year.  But it causes load on the
service, and who knows, at some point I might have to pay for it!?